Criando endpoint de compra de produto

+++ src/main/java/br/com/estoque/exceptions/ProdutoInexistente.java
@@ -1,7 +1,7 @@
 package br.com.estoque.exceptions;
 
 public class ProdutoInexistente extends RuntimeException{
-    public ProdutoInexistente(Long codigo) {
+    public ProdutoInexistente(Integer codigo) {
         super("Nao existe um produto cadastrado com o codigo " + codigo+ ".");
     }
 }
